moving of document commands to its own command
